Change to build-type: Simple
`cabal new-build` generates .ghc.environment files, so we
can run `doctest` on sources using package db there. This works on
GHC >= 8.0. In short

   cabal new-install -z doctest
   cabal new-build
   doctest src

If there aren't `-z` flag in your `cabal`, use a workaround to install
doctest:

   (cd /tmp && cabal new-install doctest)

`stack` users can `stack exec doctest src` too, as `stack exec` will
provide package environment in an environment variable.

The only downside, is that `doctest` binary have to correspond
to the GHC in use. This is not a problem in CI, but is inconvenient
locally.